refactor: move settings and token storage to commonMain
- Consolidated platform-specific `SettingsRepository` and `TokenStorage` implementations into `commonMain`
- Removed redundant `SettingsDataSource` and `TokenDataSource` classes
- Updated Koin modules to provide shared data implementations using `singleOf`
- Removed unused `materialIconsExtended` dependency and related imports
