+++ composeApp/build.gradle.kts
@@ -93,8 +93,6 @@ kotlin {
 
             implementation(libs.reorderable)
 
-            implementation(compose.materialIconsExtended)
-
             api(libs.androidx.datastore)
             api(libs.androidx.datastore.preferences)
         }
