+++ composeApp/src/commonMain/kotlin/io/dimasla4ee/shoppinglist/core/data/repository/SettingsRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package io.dimasla4ee.shoppinglist.app.data.repository
+package io.dimasla4ee.shoppinglist.core.data.repository
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
@@ -10,19 +10,21 @@ import io.dimasla4ee.shoppinglist.feature.products_screen.domain.SortMode
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 
-class SettingsRepositoryJVM(
+class SettingsRepositoryImpl(
     private val dataStore: DataStore<Preferences>
 ) : SettingsRepository {
 
     private companion object {
         val THEME_KEY = stringPreferencesKey("theme_mode")
-        private fun sortModeKey(listId: Long) = stringPreferencesKey("sort_mode_$listId")
+        fun sortModeKey(listId: Long) = stringPreferencesKey("sort_mode_$listId")
     }
 
-    override val themeMode: Flow<ThemeMode> = dataStore.data.map { preferences ->
-        val modeString = preferences[THEME_KEY] ?: "SYSTEM"
-        ThemeMode.valueOf(modeString)
-    }
+    override val themeMode: Flow<ThemeMode> =
+        dataStore.data.map { preferences ->
+            val raw = preferences[THEME_KEY]
+            runCatching { ThemeMode.valueOf(raw ?: ThemeMode.SYSTEM.name) }
+                .getOrDefault(ThemeMode.SYSTEM)
+        }
 
     override suspend fun setThemeMode(mode: ThemeMode) {
         dataStore.edit { preferences ->
@@ -30,27 +32,22 @@ class SettingsRepositoryJVM(
         }
     }
 
-    override fun observeSortMode(listId: Long): Flow<SortMode> = dataStore.data
-        .map { preferences ->
-            val key = sortModeKey(listId)
-            val value = preferences[key] ?: SortMode.CUSTOM.name
-            SortMode.valueOf(value)
+    override fun observeSortMode(listId: Long): Flow<SortMode> =
+        dataStore.data.map { preferences ->
+            val raw = preferences[sortModeKey(listId)]
+            runCatching { SortMode.valueOf(raw ?: SortMode.CUSTOM.name) }
+                .getOrDefault(SortMode.CUSTOM)
         }
 
     override suspend fun setSortMode(listId: Long, mode: SortMode) {
-        dataStore.updateData { preferences ->
-            preferences.toMutablePreferences().apply {
-                val key = sortModeKey(listId)
-                this[key] = mode.name
-            }
+        dataStore.edit { preferences ->
+            preferences[sortModeKey(listId)] = mode.name
         }
     }
 
     override suspend fun removeSortMode(listId: Long) {
-        dataStore.updateData { preferences ->
-            preferences.toMutablePreferences().apply {
-                remove(sortModeKey(listId))
-            }
+        dataStore.edit { preferences ->
+            preferences.remove(sortModeKey(listId))
         }
     }
 }
